requirements.txt
et_xmlfile==2.0.0
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
tzdata==2025.1
